fix(searchbox): do not overide searchbox content on pageshow
This behavior was introduced because of specificity in firefox for
handling page cache (called BFCache). This behavior was buggy on IE
when there is latency and the user add a query in the searchbox, it is
overriden.

After some tests, the behavior does not seem necesary anymore, so
removing it seems safe. Also the page about BFCache has been archived on
MDN and there are no other recent documentation on the subject.